restructure and use prometheus_push
