+++ src/commands/backup.rs
@@ -1,10 +1,11 @@
 //! `backup` subcommand
 
+use std::env;
 use std::path::PathBuf;
 
 use crate::{
     commands::{init::init, snapshots::fill_table},
-    config::hooks::Hooks,
+    config::{hooks::Hooks, parse_label},
     helpers::{bold_cell, bytes_size_to_string, table},
     repository::CliRepo,
     status_err, Application, RUSTIC_APP,
@@ -20,8 +21,9 @@ use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
 
 use rustic_core::{
-    BackupOptions, CommandInput, ConfigOptions, IndexedIds, KeyOptions, LocalSourceFilterOptions,
-    LocalSourceSaveOptions, ParentOptions, PathList, ProgressBars, Repository, SnapshotOptions,
+    repofile::SnapshotFile, BackupOptions, CommandInput, ConfigOptions, IndexedIds, KeyOptions,
+    LocalSourceFilterOptions, LocalSourceSaveOptions, ParentOptions, PathList, ProgressBars,
+    Repository, SnapshotOptions,
 };
 
 /// `backup` subcommand
@@ -72,24 +74,8 @@ pub struct BackupCmd {
     #[merge(strategy=conflate::bool::overwrite_false)]
     json: bool,
 
-    /// Push metrics to a Prometheus Pushgateway
-    ///
-    /// If the Pushgateway requires authentication, then the login/password can be passed in
-    /// the `RUSTIC_PUSHGATEWAY_USER` and `RUSTIC_PUSHGATEWAY_PASS` environment variables.
-    #[cfg(feature = "prometheus")]
-    #[clap(long, value_name = "PUSHGATEWAY_URL", value_hint = ValueHint::Url)]
-    #[merge(strategy=conflate::option::overwrite_none)]
-    prometheus: Option<String>,
-
-    /// Job name for the Pushgateway push
-    #[cfg(feature = "prometheus")]
-    #[clap(long, value_name = "JOB_NAME")]
-    #[merge(strategy=conflate::option::overwrite_none)]
-    prometheus_job: Option<String>,
-
     /// Additional labels to set to generated Prometheus metrics
-    #[cfg(feature = "prometheus")]
-    #[clap(long, value_name = "NAME=VALUE", value_parser = crate::prometheus::parse_label)]
+    #[clap(long, value_name = "NAME=VALUE", value_parser = parse_label)]
     #[merge(strategy=conflate::vec::append)]
     prometheus_labels: Vec<(String, String)>,
 
@@ -351,16 +337,174 @@ impl BackupCmd {
         }
 
         #[cfg(feature = "prometheus")]
-        if let Some(pushgateway_url) = self.prometheus {
-            crate::prometheus::publish_metrics(
-                &pushgateway_url,
-                self.prometheus_job.as_deref().unwrap_or("rustic"),
-                &snap,
-                self.prometheus_labels,
-            )?;
+        if config.global.is_prometheus_configured() {
+            publish_metrics(&snap, self.prometheus_labels)?;
         }
 
         info!("backup of {source} done.");
         Ok(())
     }
+}
+
+#[cfg(feature = "prometheus")]
+fn publish_metrics(snap: &SnapshotFile, user_labels: Vec<(String, String)>) -> Result<()> {
+    use prometheus::register_gauge;
+
+    let summary = snap.summary.as_ref().expect("Reaching the 'push to prometheus' point should only happen for successful backups, which must have a summary set.");
+
+    let metric_time = register_gauge!("rustic_backup_time", "Timestamp of this snapshot",)
+        .context("registering prometheus gauge")?;
+    metric_time.set(snap.time.timestamp_millis() as f64 / 1000.);
+    let metric_files_new = register_gauge!(
+        "rustic_backup_files_new",
+        "New files compared to the last (i.e. parent) snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_files_new.set(summary.files_new as f64);
+    let metric_files_changed = register_gauge!(
+        "rustic_backup_files_changed",
+        "Changed files compared to the last (i.e. parent) snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_files_changed.set(summary.files_changed as f64);
+    let metric_files_unmodified = register_gauge!(
+        "rustic_backup_files_unmodified",
+        "Unchanged files compared to the last (i.e. parent) snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_files_unmodified.set(summary.files_unmodified as f64);
+    let metric_total_files_processed = register_gauge!(
+        "rustic_backup_total_files_processed",
+        "Total processed files",
+    )
+    .context("registering prometheus gauge")?;
+    metric_total_files_processed.set(summary.total_files_processed as f64);
+    let metric_total_bytes_processed = register_gauge!(
+        "rustic_backup_total_bytes_processed",
+        "Total size of all processed files",
+    )
+    .context("registering prometheus gauge")?;
+    metric_total_bytes_processed.set(summary.total_bytes_processed as f64);
+    let metric_dirs_new = register_gauge!(
+        "rustic_backup_dirs_new",
+        "New directories compared to the last (i.e. parent) snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_dirs_new.set(summary.dirs_new as f64);
+    let metric_dirs_changed = register_gauge!(
+        "rustic_backup_dirs_changed",
+        "Changed directories compared to the last (i.e. parent) snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_dirs_changed.set(summary.dirs_changed as f64);
+    let metric_dirs_unmodified = register_gauge!(
+        "rustic_backup_dirs_unmodified",
+        "Unchanged directories compared to the last (i.e. parent) snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_dirs_unmodified.set(summary.dirs_unmodified as f64);
+    let metric_total_dirs_processed = register_gauge!(
+        "rustic_backup_total_dirs_processed",
+        "Total processed directories",
+    )
+    .context("registering prometheus gauge")?;
+    metric_total_dirs_processed.set(summary.total_dirs_processed as f64);
+    let metric_total_dirsize_processed = register_gauge!(
+        "rustic_backup_total_dirsize_processed",
+        "Total number of data blobs added by this snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_total_dirsize_processed.set(summary.total_dirsize_processed as f64);
+    let metric_data_blobs = register_gauge!(
+        "rustic_backup_data_blobs",
+        "Total size of all processed dirs",
+    )
+    .context("registering prometheus gauge")?;
+    metric_data_blobs.set(summary.data_blobs as f64);
+    let metric_tree_blobs = register_gauge!(
+        "rustic_backup_tree_blobs",
+        "Total number of tree blobs added by this snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_tree_blobs.set(summary.tree_blobs as f64);
+    let metric_data_added = register_gauge!(
+        "rustic_backup_data_added",
+        "Total uncompressed bytes added by this snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_data_added.set(summary.data_added as f64);
+    let metric_data_added_packed = register_gauge!(
+        "rustic_backup_data_added_packed",
+        "Total bytes added to the repository by this snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_data_added_packed.set(summary.data_added_packed as f64);
+    let metric_data_added_files = register_gauge!(
+        "rustic_backup_data_added_files",
+        "Total uncompressed bytes (new/changed files) added by this snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_data_added_files.set(summary.data_added_files as f64);
+    let metric_data_added_files_packed = register_gauge!(
+        "rustic_backup_data_added_files_packed",
+        "Total bytes for new/changed files added to the repository by this snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_data_added_files_packed.set(summary.data_added_files_packed as f64);
+    let metric_data_added_trees = register_gauge!(
+        "rustic_backup_data_added_trees",
+        "Total uncompressed bytes (new/changed directories) added by this snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_data_added_trees.set(summary.data_added_trees as f64);
+    let metric_data_added_trees_packed = register_gauge!(
+        "rustic_backup_data_added_trees_packed",
+        "Total bytes (new/changed directories) added to the repository by this snapshot",
+    )
+    .context("registering prometheus gauge")?;
+    metric_data_added_trees_packed.set(summary.data_added_trees_packed as f64);
+    let metric_backup_start = register_gauge!(
+        "rustic_backup_backup_start",
+        "Start time of the backup. This may differ from the snapshot `time`.",
+    )
+    .context("registering prometheus gauge")?;
+    metric_backup_start.set(summary.backup_start.timestamp_millis() as f64 / 1000.);
+    let metric_backup_end = register_gauge!(
+        "rustic_backup_backup_end",
+        "The time that the backup has been finished.",
+    )
+    .context("registering prometheus gauge")?;
+    metric_backup_end.set(summary.backup_end.timestamp_millis() as f64 / 1000.);
+    let metric_backup_duration = register_gauge!(
+        "rustic_backup_backup_duration",
+        "Total duration of the backup in seconds, i.e. the time between `backup_start` and `backup_end`",
+    ).context("registering prometheus gauge")?;
+    metric_backup_duration.set(summary.backup_duration);
+    let metric_total_duration = register_gauge!(
+        "rustic_backup_total_duration",
+        "Total duration that the rustic command ran in seconds",
+    )
+    .context("registering prometheus gauge")?;
+    metric_total_duration.set(summary.total_duration);
+
+    let mut labels = vec![
+        ("paths".to_owned(), format!("{}", snap.paths)),
+        ("hostname".to_owned(), snap.hostname.clone()),
+        ("uid".to_owned(), format!("{}", snap.uid)),
+        ("gid".to_owned(), format!("{}", snap.gid)),
+    ];
+    // See https://github.com/tikv/rust-prometheus/issues/535
+    if !snap.label.is_empty() {
+        labels.push(("snapshot_label".to_string(), snap.label.clone()));
+    }
+    if !snap.username.is_empty() {
+        labels.push(("username".to_string(), snap.username.clone()));
+    }
+    let tags = format!("{}", snap.tags);
+    if !tags.is_empty() {
+        labels.push(("tags".to_string(), tags));
+    }
+    labels.extend(user_labels);
+
+    RUSTIC_APP.config().global.push_metrics(labels)
 }
